Update code syntax and library version
The changes in this commit include updating the syntax for class and array initialization in various files including ReservedClientsModel.cs and ReserveCommand.cs. It also updates conditional checks in the Plugin.cs file, changing '.Any()' to a count comparison for improved performance. The target framework and library version is also updated in the Reservio.csproj file.